requirements.txt
click==6.7
google-cloud==0.22.0
mysqlclient==1.3.9
